SummaryView handles view for summary vc
