+++ WellnessDiary.xcodeproj/xcuserdata/luisflores.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -3,70 +3,4 @@
    uuid = "29D86646-BBE7-4F5B-BF6C-AC7604EDBC8A"
    type = "1"
    version = "2.0">
-   <Breakpoints>
-      <BreakpointProxy
-         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
-         <BreakpointContent
-            uuid = "25268566-2A82-4D42-A0C7-6AEB0DC9B47E"
-            shouldBeEnabled = "Yes"
-            ignoreCount = "0"
-            continueAfterRunningActions = "No"
-            filePath = "WellnessDiary/UI/ViewControllers/Summary/WDMSummaryTableViewController.swift"
-            startingColumnNumber = "9223372036854775807"
-            endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "282"
-            endingLineNumber = "282"
-            landmarkName = "loadChartData()"
-            landmarkType = "7">
-            <Locations>
-               <Location
-                  uuid = "25268566-2A82-4D42-A0C7-6AEB0DC9B47E - 47c53176895bd620"
-                  shouldBeEnabled = "Yes"
-                  ignoreCount = "0"
-                  continueAfterRunningActions = "No"
-                  symbolName = "WellnessDiary.WDMSummaryTableViewController.(loadChartData in _4F8DC343DB4936E63DFC09B2D6D90265)() -&gt; ()"
-                  moduleName = "WellnessDiary"
-                  usesParentBreakpointCondition = "Yes"
-                  urlString = "file:///Users/luisflores/Desktop/WellnessDiaryMobile/WellnessDiary/UI/ViewControllers/Summary/WDMSummaryTableViewController.swift"
-                  startingColumnNumber = "9223372036854775807"
-                  endingColumnNumber = "9223372036854775807"
-                  startingLineNumber = "281"
-                  endingLineNumber = "281"
-                  offsetFromSymbolStart = "6706">
-               </Location>
-               <Location
-                  uuid = "25268566-2A82-4D42-A0C7-6AEB0DC9B47E - d9aa623215d57ed8"
-                  shouldBeEnabled = "Yes"
-                  ignoreCount = "0"
-                  continueAfterRunningActions = "No"
-                  symbolName = "closure #4 (CareKitUI.OCKCartesianChartView) -&gt; () in WellnessDiary.WDMSummaryTableViewController.(loadChartData in _4F8DC343DB4936E63DFC09B2D6D90265)() -&gt; ()"
-                  moduleName = "WellnessDiary"
-                  usesParentBreakpointCondition = "Yes"
-                  urlString = "file:///Users/luisflores/Desktop/WellnessDiaryMobile/WellnessDiary/UI/ViewControllers/Summary/WDMSummaryTableViewController.swift"
-                  startingColumnNumber = "9223372036854775807"
-                  endingColumnNumber = "9223372036854775807"
-                  startingLineNumber = "282"
-                  endingLineNumber = "282"
-                  offsetFromSymbolStart = "477">
-               </Location>
-            </Locations>
-         </BreakpointContent>
-      </BreakpointProxy>
-      <BreakpointProxy
-         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
-         <BreakpointContent
-            uuid = "E4923516-9884-4D58-AFA8-31337E17ADA8"
-            shouldBeEnabled = "Yes"
-            ignoreCount = "0"
-            continueAfterRunningActions = "No"
-            filePath = "WellnessDiary/Model/CoreData/ResearchKit/ResearchKitStoreManager.swift"
-            startingColumnNumber = "9223372036854775807"
-            endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "106"
-            endingLineNumber = "106"
-            landmarkName = "addSurvey(withSleepQuantity:sleepQuality:temperature:bloodPressure:heartBeat:weight:sugarLevel:painLevel:date:persist:notify:)"
-            landmarkType = "7">
-         </BreakpointContent>
-      </BreakpointProxy>
-   </Breakpoints>
 </Bucket>
