+++ WellnessDiary/UI/Common/Views/WDMSimpleView.swift
@@ -1,0 +1,13 @@
+//
+//  WDMSimpleView.swift
+//  WellnessDiary
+//
+//  Created by Luis Flores on 4/12/21.
+//  Copyright © 2021 Health Essential Technologies. All rights reserved.
+//
+
+import UIKit
+
+class WDMSimpleView: UIView {
+  
+}
